Fix lockup in job_client, update code for python 3.5 compatibility, display progress when exporting takes

+++ job_client/job_client.py
@@ -2,6 +2,9 @@
 # Copyright (c) 2017 Electronic Arts Inc. All Rights Reserved 
 #
 
+from __future__ import print_function
+from builtins import object
+from builtins import str
 import json
 import time
 import argparse
@@ -45,12 +48,10 @@ def kill_process_tree(pid, including_parent=True):
 def class_from_string(name):
     return getattr(importlib.import_module('.'.join(name.split('.')[:-1])), name.split('.')[-1])
 
-def job_process(job_id, job_class, q, pipe, server_url, log_filename=None):
+def job_process(job_id, job_class, parameters, q, pipe, server_url, log_filename=None):
 
     logger = None
     try:
-        parameters = q.get()
-
         logger = logging.getLogger('JobProcessLogger') # TODO we may need a unique name in linux ?
         logger.setLevel(logging.DEBUG)
 
@@ -84,6 +85,7 @@ def job_process(job_id, job_class, q, pipe, server_url, log_filename=None):
         if logger:
             logger.debug("Job Yield to children")
 
+        # Place job output in the queue
         q.put({'job_id':job_id, 'waiting':True, 'retcode':0, 'children':e.children})
 
         logger.info('Job YieldToChildren: #%d' % job_id)
@@ -103,7 +105,7 @@ def job_process(job_id, job_class, q, pipe, server_url, log_filename=None):
 
     pipe.close()
 
-class JobInstance():
+class JobInstance(object):
 
     def __str__(self):
         return 'Job #%d %s' % (self.job_id, self.status)
@@ -121,7 +123,7 @@ class JobInstance():
         self.terminated = False
 
         try:
-            if isinstance(job_class, basestring):
+            if isinstance(job_class, str) or isinstance(job_class.decode(), str):
                 self.job_class = class_from_string(job_class)
 
             if not self.job_class.__module__.startswith('jobs.'):
@@ -135,43 +137,56 @@ class JobInstance():
             exception_details = traceback.format_exc()
             self.result = {'job_id':self.job_id, 'success':False, 'retcode':1, 'exception':exception_details, 'progress':self.status}
     
-    def output_recieved_from_job(self, data):
-        print 'Job #%d> %s' % (self.job_id, data)
+    def output_received_from_job(self, data):
+        print('Job #%d> %s' % (self.job_id, data))
         self.status = data
 
     def terminate(self):
-        print 'Terminate Job #%d' % (self.job_id)
+        print('Terminate Job #%d' % (self.job_id))
         try:
             if self.p:
+                if not self.p.is_alive():
+                    return
                 self.terminated = True
                 kill_process_tree(self.p.pid, True)
         except Exception as e:
-            print 'Could not terminate job #%d : %s' % (self.job_id, e)
+            print('Could not terminate job #%d : %s' % (self.job_id, e))
 
     def blocking_run(self):
         parent_conn, child_conn = Pipe()
         q = Queue()
-        q.put(self.parameters)
-        self.p = Process(target=job_process, args=(self.job_id, self.job_class, q, child_conn, self.server_url, self.log_filename, ))
+        self.p = Process(target=job_process, args=(self.job_id, self.job_class, self.parameters, q, child_conn, self.server_url, self.log_filename, ))
         self.p.start()
         while self.p.is_alive():
             while parent_conn.poll():
-                self.output_recieved_from_job(parent_conn.recv())
+                self.output_received_from_job(parent_conn.recv())
             time.sleep(1)
         self.p.join()
-        while parent_conn.poll():
-            self.output_recieved_from_job(parent_conn.recv())
+        try:
+            while parent_conn.poll():
+                self.output_received_from_job(parent_conn.recv())
+        except:
+            print('Exception while gathering job process output')
 
-        if self.terminated:
-            self.result = {'job_id':self.job_id, 'success':False, 'retcode':1, 'exception':'Terminated by server', 'progress':'terminated'}
-        else:
-            self.result = q.get()
-            self.result['progress'] = self.status
+        try:
+            if self.terminated:
+                # Job was terminated from user request
+                self.result = {'job_id':self.job_id, 'success':False, 'retcode':1, 'exception':'Terminated by server', 'progress':'terminated'}
+            else:
+                if q.empty():
+                    # Job process error, the job should always place its output in the queue
+                    self.result = {'job_id':self.job_id, 'success':False, 'retcode':1, 'exception':'Job process terminated abnormally', 'progress':'failed'}
+                else:                
+                    # Job terminated normally
+                    self.result = q.get()
+                    self.result['progress'] = self.status
+        except:
+            print('Exception while gathering job output')
 
-        parent_conn.close()    
+        parent_conn.close()   
         return self.result
 
-class JobContainer():
+class JobContainer(object):
     def __init__(self, log_folder, server_url):
         self.log_folder = log_folder
         self.server_url = server_url
@@ -215,17 +230,17 @@ class JobContainer():
 
     def wait(self):
         ''' Wait for all current jobs to be finished '''
-        for t,j in self.running_jobs.values():
+        for t,j in list(self.running_jobs.values()):
             t.join()
 
-class JobClient():
+class JobClient(object):
 
     LOG_FOLDER = os.path.abspath('logs')
     PHONEHOME_DELAY = 4 # seconds
 
-    def __init__(self, server):
+    def __init__(self, server, git_version):
 
-        print 'Logs will be written to %s' % self.LOG_FOLDER
+        print('Logs will be written to %s' % self.LOG_FOLDER)
 
         self.USERNAME = DEFAULT_USERNAME
         self.PASSWORD = DEFAULT_PASSWORD
@@ -244,6 +259,7 @@ class JobClient():
         self.first_update = True
         self.need_restart = False
         self.cpu_percent = 0.0
+        self.git_version = git_version
 
         # Log File Logger
         self.logger = logging.getLogger('JobClientLogger')
@@ -263,10 +279,22 @@ class JobClient():
         try:  
             import vhtrack
             self.cuda_device_count = vhtrack.query_cuda_device_count()
+            if self.cuda_device_count > 0:
+                di = vhtrack.DeviceInfo(0)
+                if di.isCompatible():
+                    total_memory = di.totalMemory()
+                    vram_in_GB = total_memory >> 30
+                    if vram_in_GB < 4 or di.majorVersion() < 5:
+                        self.cuda_device_count = 0
+
+                    self.logger.info('%d CUDA GPU Detected %s with %d GB VRAM' % (self.cuda_device_count, di.name(), vram_in_GB))
+                else:
+                    self.cuda_device_count = 0
         except:
             self.cuda_device_count = 0
         
-        self.logger.info('%d CUDA GPU Detected' % self.cuda_device_count)
+        if self.cuda_device_count == 0:
+            self.logger.info('No CUDA GPU Detected')
 
         self.available_job_classes = create_available_job_list(self.logger)
 
@@ -286,16 +314,17 @@ class JobClient():
             'status': 'accepting',
             'restarted': self.first_update,
             'code_version': VERSION,
-            'running_jobs': self.container.running_jobs.keys(),
+            'running_jobs': list(self.container.running_jobs.keys()),
             'running_jobs_progress': [(key,self.container.running_jobs[key][1].status) for key in self.container.running_jobs],
-            'finished_jobs': [j.result for j in self.container.finished_jobs.values()],
+            'finished_jobs': [j.result for j in list(self.container.finished_jobs.values())],
             'system' : platform.system(),
             'system_bits' : self.info['bits'],
             'cpu_brand' : self.info['brand'],
             'cpu_cores' : self.info['count'],
             'cuda_device_count' : self.cuda_device_count,
             'available_jobs' : self.available_job_classes,
-            'cpu_percent' : self.cpu_percent
+            'cpu_percent' : self.cpu_percent,
+            'git_version' : self.git_version
             }
 
         for j in payload['finished_jobs']:
@@ -310,7 +339,7 @@ class JobClient():
             for x in payload['finished_jobs']:
                 del self.container.finished_jobs[x['job_id']]
 
-            # Launch new jobs we just recieved
+            # Launch new jobs we just received
             try:
                 data = r.json()
 
@@ -321,7 +350,7 @@ class JobClient():
 
                 if 'jobs' in data:
                     for new_job in data['jobs']:
-                        self.logger.info('Job recieved: #%d %s' % (new_job['job_id'], new_job['job_class']))
+                        self.logger.info('Job received: #%d %s' % (new_job['job_id'], new_job['job_class']))
                         self.container.submit_job(**new_job)
                         self.status_changed = True
 
@@ -330,7 +359,7 @@ class JobClient():
                         self.container.terminate_job(job_id)                        
 
             except Exception as e:
-                self.logger.error('Recieved invalid data from server. %s' % e)
+                self.logger.error('received invalid data from server. %s' % e)
         else:
             self.logger.error("Unexpected return code while contacting server %s (%d)" % (self.server, r.status_code))
             self.logger.error(r.text)
@@ -346,7 +375,7 @@ class JobClient():
                     self.logger.error('Failed to contact server. %s' % e)
                 
                 if self.need_restart and not self.container.running_jobs and not self.container.finished_jobs:
-                    print 'Restarting...'
+                    print('Restarting...')
                     # End current process and launch a new one with the same command line parameters
                     #os.execvp('python.exe', ['python.exe'] + sys.argv)
                     exit(3)
@@ -357,25 +386,51 @@ class JobClient():
         except KeyboardInterrupt:
             pass
         finally:
-            print 'Exiting...'
+            print('Exiting...')
             self.notify_exit()
 
-class SingleInstance:
+class SingleInstance(object):
     def __init__(self, name):
-        self.lock_file = os.path.join(os.path.split(sys.modules[__name__].__file__)[0], name+'.lock')
-        self.fd = None
-        try:
-            if os.path.exists(self.lock_file):
-                os.unlink(self.lock_file)
-            self.fd = os.open(self.lock_file, os.O_CREAT | os.O_EXCL | os.O_RDWR)
-        except OSError:
-            type, e, tb = sys.exc_info()
-            if e.errno == 13:
+        self.locked = False
+        self.lock_file = os.path.abspath(os.path.join(os.path.split(sys.modules[__name__].__file__)[0], name+'.lock'))
+        if sys.platform == 'win32':
+            # Windows
+            self.fd = None
+            try:
+                if os.path.exists(self.lock_file):
+                    os.unlink(self.lock_file)
+                self.fd = os.open(self.lock_file, os.O_CREAT | os.O_EXCL | os.O_RDWR)
+                self.locked = True
+            except OSError:
+                type, e, tb = sys.exc_info()
+                if e.errno == 13:
+                    raise Exception('Process %s Already Running' % name)
+        else:
+            # Linux
+            import fcntl
+            self.fp = open(self.lock_file, 'w')
+            self.fp.flush()
+            try:
+                fcntl.lockf(self.fp, fcntl.LOCK_EX | fcntl.LOCK_NB)
+                self.locked = True
+            except IOError:
                 raise Exception('Process %s Already Running' % name)
+
+
     def __del__(self):
-        if self.fd:
-            os.close(self.fd)
-            os.unlink(self.lock_file)
+        if self.locked:
+            if sys.platform == 'win32':
+                # Windows
+                if self.fd:
+                    os.close(self.fd)
+                    os.unlink(self.lock_file)
+            else:
+                # Linux
+                import fcntl
+                fcntl.lockf(self.fp, fcntl.LOCK_UN)
+                # os.close(self.fp)
+                if os.path.isfile(self.lock_file):
+                    os.unlink(self.lock_file)        
 
 def create_available_job_list(logger):            
     # Make list of available Job classes
@@ -394,10 +449,13 @@ def create_available_job_list(logger):
         try:
             job = c()
             job.check_requirements()
-            job_class_names.append(str(c))
-            print ' OK %s' % c
+            class_name = str(c)
+            if "'" in class_name:
+                class_name = class_name.split("'")[1]
+            job_class_names.append(class_name)
+            print(' OK %s' % c)
         except Exception as e:
-            print 'FAIL! %s' % c        
+            print(' Not Available: %s (%s)' % (c,e.message))
     return job_class_names                   
 
 if __name__ == "__main__":
@@ -406,11 +464,12 @@ if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description='Launch Job Client')
     parser.add_argument('server', nargs='?', type=str, help='Http URL of Server, default is %s' % DEFAULT_SERVER, default=DEFAULT_SERVER)
+    parser.add_argument('--git-version', dest='version', nargs='?', type=str, help='Git Version description', default='unknown')
 
     args = parser.parse_args()
 
-    print 'Job Client Running (%s) ... Press Ctrl-C to stop' % args.server
+    print('Job Client Running (%s) ... Press Ctrl-C to stop' % args.server)
 
-    client = JobClient(args.server)
+    client = JobClient(args.server, args.version)
     client.loop_forever()
